pyproject.toml
[tool.poetry]
name = "simple-social"
version = "0.1.0"
description = ""
authors = ["Prashant Mishra <prashant.mishra@tophire.com>"]
readme = "README.md"
packages = [{include = "simple_social"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.7"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.4"
django-neomodel = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

